chore: move functions to load.go
